Rework the interface design of linksrc to include fewer exported identifiers.
This commit un-exports as many identifiers in linksrc as possible. This also meant
finding a better design for `func newSet(r io.Reader, c config) (Set, error)`, where
`config` is the struct containing package-specific types like cascadia.Selector. Since
the config struct needs to be unexported, newSet has to be called internally. But
since newSet isn't called internally, it didn't make sense to leave it unexported, or
to call it with the unexported config object. The solution is to change the interface
to `NewSet(r io.Reader, c Config)`, and validate user configs before parsing the HTML.

Note that we'll eventually need to use some kind of config object to know where to
send HTTP requests to get the HTML to parse, so we'll also want to validate the URL for
a separate package related to sending the requests.