Send an email recognized by a mail server
Get an e2e test to pass that collects sent messages from
a local SMTP server. This required making some other
changes:

- Replaced the gomail package with the standard library smtp
  package, since the former is no longer maintained.
- For simplicity in the MVP, enforce local SMTP servers and
  don't implement TLS.
- Use the yaml package for unmarshaling yaml.
- Add custom unmarshaler for poller.Config.
- Allow the app to handle interrupts.
- In the e2e test, read from the application's stderr pipe
in the background. This way, reads won't be affected when
exiting the application closes the read end of the pipe. And we
won't block indefinitely + time out if no data is available on
the write end of the pipe.
- Add info logging to main.go
- Remove shared memory from the link source pollers. These
goroutines used to refer to the db and emailData variables
without using channels, which is bad practice.
- Remove the need to serialize a linksrc.Set in order to get its
  key.
- Overhaul the way we determine LinkItem newness. Previously,
 Set.NewSince was intended to determine which LinkItems in one
Set were absent in another Set by using sort.Search. However,
 sort.Search assumes that any index in a slice greater than the one
that matches the search will also match the search.
This meant that we'd have to rethink Set.NewSince. Either we'd have
to implement an efficient search algorithm or implement a linear
search over two slices and potentially have O(n^2). However, it
turns out that the giant hash map that BadgerDB provides is
already perfect for our searching needs. We could check whether a
given LinkItem was already present in the database by using the
hash of it as a key. This change implements that new search
method.
- add a for loop around the select statement in main.go, which
stops the main goroutine from deadlocking.
- remove some channels that both send to and receive from the
  main goroutine, which is unnecessary and confusing (and causes
  deadlocks)
- Rewrite SendNewsletter to create a multipart body manually