Expose SMTP AUTH credentials as config options
Previously, we supported either a local SMTP relay with no
authentication or SendGrid's SMTP API. The latter turned out
not to be appropriate for this project, since the free tier
requires a lot of verification of what emails will look
like and how users will opt in/out. Since all major SMTP
APIs require a username/password, the most flexible way to
handle whatever SMTP relay a user sends the newsletter to
is to expose the username/password as config options.

- Use plain AUTH with all email sending
- Enforce STARTTLS for plain AUTH
- Move test SMTP server-related code to a shared package
- Replace MailHog with an in-process SMTP server that
  supports AUTH and STARTTLS
- Allow an option to skip host verification in TLS
  connections (for testing)