Simplify database cleanup
Previously, the TTL for database keys was exposed as a user
config option. However, this option only really needs to
be more than one polling interval, so keys don't expire
before we check one round of links against the database.

This removes the keyTTL config option and sets the TTL of
new database keys to double the polling interval. This means
that one round of links can be checked against a new round
of links before it's removed from the database.

Also:
- remove cleanupInterval references
- move all user config unmarshalling into
  userconfig.Meta.UnmarshalYAML
- add a scraping config section that includes the polling
  interval and data directory, which was necessary for a
  new BadgerDB to know about the polling interval.
- Improves the e2e test for cleanup. Measures the results
  of the cleanup test using standard deviation rather than
  variance, since it's in the same unit as the file sizes.
- Removes the max entries setting from Badger, since it doesn't
  affect the results of the cleanup test. What does is the setting
  to compact on close.