Make errors visible via email
This application was designed to run with minimal operator
involvement. However, when an error takes place in scraping
a website, the only way to know is to access logs on the
local machine.

At the same time, because the error channel
was unbuffered, a scraper sending to the channel would
block until the main goroutine read from the channel. But
because the main goroutine waits for all scrapers to finish
scraping before it reads from the error channel, there would
be a deadlock.

This commit changes the scraping logic to add scraping-related
errors to email messages, meaning that (a) the main goroutine
doesn't need to receive errors via an error channel and (b)
users can stay informed about scraping errors without needing
to SSH into their One Newsletter hosts.

- Don't return errors in NewSet and email generation. To do
  this, add a messages property to Set, plus methods to
  add/get messages without having direct access to the messages
  slice. Then add messages directly to an email message rather
  than using a Status enum.

- Replace the use of linksrc.Set in EmailData. Use
  BodySectionContent, so (a) we don't need to access a Set
  directly for testing and (b) so we don't need to repeat work/
  allocations in order to add LinkItems to an email body.

- Refactor email template execution. Use one function between
  GenerateText and Generate Body.