Change caption extraction behavior
Currently, One Newsletter extracts captions by finding the block-level
element with the longest combined text nodes. The intention here was to
pick the most descriptive text for a link item without including less
helpful elements like bylines and tags.

However, this approach has a drawback that can make a caption
incomprehensible: some short block-level elements are necessary to put
longer ones in context. An example of this is when an article's
headline, which the current approach often omits, introduces names that
the longest sentence in the caption refers to.

Another example is when the text in a link item does not use complete
sentences, so the role for the text nodes that make it into the caption
often differ: sometimes a byline goes missing, sometimes a title does,
etc.

This change uses the first 20 words of a link item's text for a caption.
To remove meaningless text, it removes all block-level text nodes with
word lengths that fall below a user-defined threshold, which defaults to
3. If a caption has more than 20 usable words, One Newsletter truncates
the caption with an elipsis. One Newsletter adds a period to separate
text nodes from different block-level elements.

This change also skips saving or sending a link item when it includes an
empty caption.