Fix link extraction with mixed hostnames
In link sources that include links to the same site as well as external
sites, the scraper would treat all links as though they targeted the
same site.

This is because it uses the hostname of a configured link source to
build the URLs to include in the emails it sends. If a link URL has a
hostname that differs from the link source's hostname, the scraper uses
the link source's hostname anyway.

This change fixes this behavior and uses the hostname found within a
link URL if it is available.