Add RSS and Atom feed support
When autodetecting link items from a link source URL, first check
whether the page source indicates an RSS feed, Atom feed, or HTML page,
then proceed with the autodetection.

From there, parsing RSS and Atom feeds is straightforward--we use a
library--and so is translating the library's feed representation into a
`[]LinkItem`.