Add support for notification schedules
Rather than send the newsletter at user-defined intervals, expose
notification schedules in the configuration, letting the user configure
the newsletter to be sent at specific days and times. The scraping
configuration now takes a schedule store, which contains configured
notification schedules.

Every tick of a ticker, we look up the current time within the schedule
store. If there is a match, we send a newsletter. Notification schedules
are currently limited to hours and days of the week: there cannot be
multiple mailings of the same newsletter within the same hour.

Also remove the `IterationLimit` configuration field, since this was
used in e2e tests to guarantee a certain number of email sends. But this
logic doens't make sense if we're using notification schedules.